move MSALCredentialMethod to internal folder.
