+++ lib/native/MSALNativeCredManagment/MSALNativeCredManagment/src/public/credential_method/MSALCredentialMethodProtocol.swift
@@ -1,0 +1,49 @@
+//
+// Copyright (c) Microsoft Corporation.
+// All rights reserved.
+//
+// This code is licensed under the MIT License.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files(the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions :
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+import Foundation
+
+// MARK: - Protocol
+
+/// Protocol defining the contract for all credential method types.
+///
+/// All credential method classes must conform to this protocol.
+/// Use this protocol when you need to work with credential methods generically.
+public protocol MSALCredentialMethodProtocol: AnyObject {
+
+    /// Unique identifier of the credential method.
+    var id: String { get }
+
+    /// The type identifier string (e.g., "passkey", "phone", "password").
+    var credentialType: String { get }
+
+    /// Display-friendly name or hint (e.g., masked phone "+1 ***-***-1234").
+    var displayName: String? { get }
+
+    /// Whether this is the default/primary method.
+    var isDefault: Bool { get }
+
+    /// Timestamp of when this method was registered.
+    var createdAt: Date? { get }
+}
